# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MYE030 Data Integration & Visualization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/years": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Summaries",
        "description": "Return the year summary timeseries optionally bounded by a range.\n\nThe last year of the corpus is sometimes a partial export (DBLP\nsnapshots taken mid-year). It is trimmed by default; pass\n``include_partial_last_year=true`` to receive every row.",
        "operationId": "list_year_summaries_api_years_get",
        "parameters": [
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "include_partial_last_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Partial Last Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/YearSummary"
                  },
                  "title": "Response List Year Summaries Api Years Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/years/{year}": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "Get Year Summary",
        "description": "Return the aggregate summary for one publication year.",
        "operationId": "get_year_summary_api_years__year__get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YearSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/years/{year}/journals": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Journals",
        "description": "Return journals that published in the requested year.\n\nSortable columns: ``title``, ``publisher``, ``best_quartile``,\n``sjr_index``, ``articles_in_year``.",
        "operationId": "list_year_journals_api_years__year__journals_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_YearJournalEntry_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/years/{year}/articles": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Articles",
        "description": "Return every article published in the year, optionally filtered.\n\nSortable columns: ``title``, ``venue_title``, ``venue_type``.",
        "operationId": "list_year_articles_api_years__year__articles_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "conference_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conference Id"
            }
          },
          {
            "name": "journal_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Journal Id"
            }
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Author Id"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_YearArticle_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/years/{year}/conferences": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Conferences",
        "description": "Return conferences that held papers in the requested year.\n\nSortable columns: ``title``, ``acronym``, ``rank_value``,\n``articles_in_year``.",
        "operationId": "list_year_conferences_api_years__year__conferences_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_YearConferenceEntry_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_YearArticle_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/YearArticle"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[YearArticle]"
      },
      "PaginatedResponse_YearConferenceEntry_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/YearConferenceEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[YearConferenceEntry]"
      },
      "PaginatedResponse_YearJournalEntry_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/YearJournalEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[YearJournalEntry]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YearArticle": {
        "properties": {
          "article_id": {
            "type": "integer",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "venue_type": {
            "type": "string",
            "title": "Venue Type"
          },
          "venue_id": {
            "type": "integer",
            "title": "Venue Id"
          },
          "venue_title": {
            "type": "string",
            "title": "Venue Title"
          },
          "pages": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "venue_type",
          "venue_id",
          "venue_title"
        ],
        "title": "YearArticle",
        "description": "Article published in a given year, regardless of venue type."
      },
      "YearConferenceEntry": {
        "properties": {
          "conference_id": {
            "type": "integer",
            "title": "Conference Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym"
          },
          "rank_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Value"
          },
          "primary_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For"
          },
          "primary_for_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For Description"
          },
          "articles_in_year": {
            "type": "integer",
            "title": "Articles In Year"
          }
        },
        "type": "object",
        "required": [
          "conference_id",
          "title",
          "articles_in_year"
        ],
        "title": "YearConferenceEntry",
        "description": "Conference that held at least one paper in the requested year."
      },
      "YearJournalEntry": {
        "properties": {
          "journal_id": {
            "type": "integer",
            "title": "Journal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "best_quartile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Quartile"
          },
          "best_subject_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Subject Area"
          },
          "sjr_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sjr Index"
          },
          "articles_in_year": {
            "type": "integer",
            "title": "Articles In Year"
          }
        },
        "type": "object",
        "required": [
          "journal_id",
          "title",
          "articles_in_year"
        ],
        "title": "YearJournalEntry",
        "description": "Journal that published at least one article in the requested year."
      },
      "YearSummary": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "journal_articles": {
            "type": "integer",
            "title": "Journal Articles"
          },
          "conference_articles": {
            "type": "integer",
            "title": "Conference Articles"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles"
          },
          "distinct_journals": {
            "type": "integer",
            "title": "Distinct Journals"
          },
          "distinct_conferences": {
            "type": "integer",
            "title": "Distinct Conferences"
          },
          "distinct_authors": {
            "type": "integer",
            "title": "Distinct Authors"
          },
          "total_authors": {
            "type": "integer",
            "title": "Total Authors"
          }
        },
        "type": "object",
        "required": [
          "year",
          "journal_articles",
          "conference_articles",
          "total_articles",
          "distinct_journals",
          "distinct_conferences",
          "distinct_authors",
          "total_authors"
        ],
        "title": "YearSummary",
        "description": "Aggregate counts for a single publication year."
      }
    }
  }
}